+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+This version adds support for Omeka S 4.0; the minimum supported version of
+Omeka S is 3.0.
+
 ## [0.3.2] - 2022-02-18
 
 ### Fixed